requirements.txt
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0
nltk>=3.8.0
textblob>=0.18.0
vaderSentiment>=3.3.2
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0
plotly>=5.18.0
gensim>=4.3.0
sqlalchemy>=2.0.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0
emoji>=2.8.0
regex>=2023.10.0
